feat: complete route module migration to TypeScript
Migrate remaining 3 route modules to TypeScript:
- upload.ts (556 lines) - file upload, AI generation
- art.ts (776 lines) - search, smart-search, similar, random, import
- devices.ts (886 lines) - device status, commands, battery tracking

All 12 route modules now fully migrated. 220 tests passing.
